Introduce support for TypeScript files using @babel/preset-typescript
This introduces support for TypeScript files in the project. Note that does not mean that TypeScript files are compiled, rather they are stripped of their types before compilation. This will allow us to migrate our code slowly to be more strictly typed and eventually use the TypeScript compiler as well.